(libfreefare-desfire) Fix the mifare_desfire_key_set_version() function.
The documentation does not say if the version bits should be stored in the
higher or lower part of a 3DES key not bits position.  So for now, make
mifare_desfire_key_get_version() and mifare_desfire_key_set_version()
consistent.  The regression test suite is due to test on-card key modifications
so these functions validity will be enforcer in a near future.