(libfreefare-desfire) Fix key version storage.
As detailed int the r424 commit log, it was just a matter of time to know which
of the get or set_key_version was the good one (I could not find this
information in the documentation).  The updated regression test suite that is
WIP and will be committed soon confirms which of the two functions was right:
another proof of Murphy's laws: I changed the wrong function to match the other
one in r424.