(libfreefare-desfire) Implement DES/3DES encryption for writing data.
Okay, Housto^WSaint-Quentin, we've had a problem here.
The idea was to make the API hardly user-friendly and have the library detect
what communication settings where to use (plain vs. DES/3DES encryption vs.
MACing).  However, the mifare_desfire_get_file_settings() function may be
allowed only when authenticated with the application master key.  This makes it
impossible to setup a reliable and trivial mode selection feature.  For now,
assume that file settings can be fetched, in the future however it will be
required to choose on of the following:
  - Always ask the user the communication settings to use;
  - Provide two functions, one with auto-detection feature and another with
    user-provided communication settings;
  - Try to send a message with all modes in turn until one works-out.